feat: classes & instances, set/get properties

+++ include/native_functions.h
@@ -2,13 +2,13 @@
 #include "object.h"
 #include "value.h"
 
-value_t gustav_clock_native(size_t arg_count [[maybe_unused]],
+value_t gustav_clock_native(int arg_count [[maybe_unused]],
 			    value_t *args [[maybe_unused]]);
 
-value_t gustav_sleep_native(size_t arg_count [[maybe_unused]],
+value_t gustav_sleep_native(int arg_count [[maybe_unused]],
 			    value_t *args [[maybe_unused]]);
 
-value_t gustav_max_native(size_t arg_count [[maybe_unused]],
+value_t gustav_max_native(int arg_count [[maybe_unused]],
 			  value_t *args [[maybe_unused]]);
 
 typedef struct {
